\name{SNPmaxsel-internal}
\alias{chisq.test.a}
\alias{groups}
\alias{makeA}
\alias{makeB}


\title{Internal SNPmaxsel Functions}
\description{
  Internal SNPmaxsel functions.
}
\note{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written).
}
\keyword{internal}
